feat(category): :sparkles: handle image url in category
