+++ backend/api/routes/categoryRoutes.js
@@ -1,4 +1,3 @@
-// routes/categoryRoutes.js
 const express = require('express');
 const router = express.Router();
 const categoryController = require('../controllers/categoryController');
@@ -19,9 +18,29 @@ const categoryController = require('../controllers/categoryController');
  *               name:
  *                 type: string
  *                 description: Name of the category
+ *               image_url:
+ *                 type: string
+ *                 format: uri
+ *                 description: URL of the category image
+ *                 example: "https://s3.amazonaws.com/my-bucket/images/category-123.jpg"
  *     responses:
  *       201:
  *         description: Category created successfully
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 id:
+ *                   type: string
+ *                   description: Category ID
+ *                 name:
+ *                   type: string
+ *                   description: Category name
+ *                 image_url:
+ *                   type: string
+ *                   format: uri
+ *                   description: URL of the category image
  *       500:
  *         description: Error during category creation
  */
@@ -49,6 +68,10 @@ router.post('/categories', categoryController.createCategory);
  *                   name:
  *                     type: string
  *                     description: Category's name
+ *                   image_url:
+ *                     type: string
+ *                     format: uri
+ *                     description: URL of the category image
  *                   subcategories:
  *                     type: array
  *                     description: List of subcategories under the category
@@ -80,6 +103,21 @@ router.get('/categories', categoryController.getCategories);
  *     responses:
  *       200:
  *         description: Category found successfully
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 id:
+ *                   type: string
+ *                   description: Category ID
+ *                 name:
+ *                   type: string
+ *                   description: Category name
+ *                 image_url:
+ *                   type: string
+ *                   format: uri
+ *                   description: URL of the category image
  *       404:
  *         description: Category not found
  *       500:
@@ -110,9 +148,29 @@ router.get('/categories/:category_id', categoryController.getCategoryById);
  *               name:
  *                 type: string
  *                 description: The new name of the category
+ *               image_url:
+ *                 type: string
+ *                 format: uri
+ *                 description: New URL of the category image
+ *                 example: "https://s3.amazonaws.com/my-bucket/images/category-123-new.jpg"
  *     responses:
  *       200:
  *         description: Category updated successfully
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 id:
+ *                   type: string
+ *                   description: Category ID
+ *                 name:
+ *                   type: string
+ *                   description: Category name
+ *                 image_url:
+ *                   type: string
+ *                   format: uri
+ *                   description: Updated URL of the category image
  *       404:
  *         description: Category not found
  *       500:
